Refactor how we process GH Test suite board
